only report instrumentations that were loaded
Current behavior is to put the list of all the instrumentations that
exist or are included in an environment variable into a meta field. It
might make more sense to only include integrations that were
successfully loaded. That way, if someone is running, say, a simple
sinatra app, they'll see only the sinatra integration loaded and not all
of the available integrations.